Extract helper functions related to E2E CSV testing
We extract these from `decisions/admin/index.spec.ts` into the shared `command.ts`, as we will want to use these for testing single dataset downloads
