+++ cypress/integration/admin/decisions/index.spec.ts
@@ -1,16 +1,4 @@
 import { format } from 'date-fns';
-import { parse } from 'csv-parse/dist/esm/sync';
-import path from 'path';
-
-type SeedDecisionDataset = {
-  profession: string;
-  organisation: string;
-  year: number;
-  status: string;
-  routes: {
-    countries: [];
-  }[];
-};
 
 describe('Listing decision datasets', () => {
   context('When I am logged in as editor', () => {
@@ -20,37 +8,30 @@ describe('Listing decision datasets', () => {
     });
 
     it('Lists all decision datasets', () => {
-      cy.readFile('./seeds/test/decision-datasets.json').then(
-        (datasets: SeedDecisionDataset[]) => {
-          const datasetsToShow = getDisplayedDatasets(datasets);
+      cy.getDisplayedDatasets().then((datasets) => {
+        cy.translate('decisions.admin.dashboard.search.foundPlural', {
+          count: datasets.length,
+        }).then((foundText) => {
+          cy.get('body').should('contain', foundText);
+        });
 
-          cy.translate('decisions.admin.dashboard.search.foundPlural', {
-            count: datasetsToShow.length,
-          }).then((foundText) => {
-            cy.get('body').should('contain', foundText);
-          });
+        datasets.forEach((dataset, index) => {
+          cy.get('tbody tr')
+            .eq(index)
+            .then(($row) => {
+              cy.wrap($row).should('contain', dataset.profession);
+              cy.wrap($row).should('contain', dataset.organisation);
+              cy.wrap($row).should('contain', dataset.year.toString());
 
-          datasetsToShow.forEach((dataset, index) => {
-            cy.get('tbody tr')
-              .eq(index)
-              .then(($row) => {
-                cy.wrap($row).should('contain', dataset.profession);
-                cy.wrap($row).should('contain', dataset.organisation);
-                cy.wrap($row).should('contain', dataset.year.toString());
+              cy.get('[data-cy=changed-by-text]').should('not.exist');
+              cy.wrap($row).should('contain', format(new Date(), 'd MMM yyyy'));
 
-                cy.get('[data-cy=changed-by-text]').should('not.exist');
-                cy.wrap($row).should(
-                  'contain',
-                  format(new Date(), 'd MMM yyyy'),
-                );
-
-                cy.translate(`app.status.${dataset.status}`).then((status) => {
-                  cy.wrap($row).should('contain', status);
-                });
+              cy.translate(`app.status.${dataset.status}`).then((status) => {
+                cy.wrap($row).should('contain', status);
               });
-          });
-        },
-      );
+            });
+        });
+      });
 
       checkCsvDownload(() => true);
     });
@@ -200,17 +181,13 @@ describe('Listing decision datasets', () => {
           .should('not.be.checked');
       });
 
-      cy.readFile('./seeds/test/decision-datasets.json').then(
-        (datasets: SeedDecisionDataset[]) => {
-          const datasetsToShow = getDisplayedDatasets(datasets);
-
-          cy.translate('decisions.admin.dashboard.search.foundPlural', {
-            count: datasetsToShow.length,
-          }).then((foundText) => {
-            cy.get('body').should('contain', foundText);
-          });
-        },
-      );
+      cy.getDisplayedDatasets().then((datasets) => {
+        cy.translate('decisions.admin.dashboard.search.foundPlural', {
+          count: datasets.length,
+        }).then((foundText) => {
+          cy.get('body').should('contain', foundText);
+        });
+      });
 
       checkCsvDownload(() => true);
     });
@@ -223,44 +200,39 @@ describe('Listing decision datasets', () => {
     });
 
     it('Lists decision datasets for my organisation', () => {
-      cy.readFile('./seeds/test/decision-datasets.json').then(
-        (datasets: SeedDecisionDataset[]) => {
-          const datasetsToShow = getDisplayedDatasets(datasets).filter(
-            (dataset) => dataset.organisation === 'Department for Education',
-          );
+      cy.getDisplayedDatasets().then((datasets) => {
+        datasets = datasets.filter(
+          (dataset) => dataset.organisation === 'Department for Education',
+        );
 
-          cy.translate(
-            `decisions.admin.dashboard.search.${
-              datasetsToShow.length > 1 ? 'foundPlural' : 'foundSingular'
-            }`,
-            {
-              count: datasetsToShow.length,
-            },
-          ).then((foundText) => {
-            cy.get('body').should('contain', foundText);
-          });
+        cy.translate(
+          `decisions.admin.dashboard.search.${
+            datasets.length > 1 ? 'foundPlural' : 'foundSingular'
+          }`,
+          {
+            count: datasets.length,
+          },
+        ).then((foundText) => {
+          cy.get('body').should('contain', foundText);
+        });
 
-          datasetsToShow.forEach((dataset, index) => {
-            cy.get('tbody tr')
-              .eq(index)
-              .then(($row) => {
-                cy.wrap($row).should('contain', dataset.profession);
-                cy.wrap($row).should('not.contain', dataset.organisation);
-                cy.wrap($row).should('contain', dataset.year.toString());
+        datasets.forEach((dataset, index) => {
+          cy.get('tbody tr')
+            .eq(index)
+            .then(($row) => {
+              cy.wrap($row).should('contain', dataset.profession);
+              cy.wrap($row).should('not.contain', dataset.organisation);
+              cy.wrap($row).should('contain', dataset.year.toString());
 
-                cy.get('[data-cy=changed-by-text]').should('not.exist');
-                cy.wrap($row).should(
-                  'contain',
-                  format(new Date(), 'd MMM yyyy'),
-                );
+              cy.get('[data-cy=changed-by-text]').should('not.exist');
+              cy.wrap($row).should('contain', format(new Date(), 'd MMM yyyy'));
 
-                cy.translate(`app.status.${dataset.status}`).then((status) => {
-                  cy.wrap($row).should('contain', status);
-                });
+              cy.translate(`app.status.${dataset.status}`).then((status) => {
+                cy.wrap($row).should('contain', status);
               });
-          });
-        },
-      );
+            });
+        });
+      });
 
       checkCsvDownload(
         (dataset) => dataset.organisation === 'Department for Education',
@@ -331,74 +303,10 @@ function checkCsvDownload(
   filter: (dataset: SeedDecisionDataset) => boolean,
 ): void {
   cy.translate('decisions.admin.dashboard.download').then((download) => {
-    clickDownloadLink(download);
-  });
-
-  const filename = path.join(
-    Cypress.config('downloadsFolder'),
-    `decisions-${format(new Date(), 'yyyyMMdd')}.csv`,
-  );
-
-  cy.readFile(filename).then((file) => {
-    const rows: string[][] = parse(file);
-
-    cy.readFile('./seeds/test/decision-datasets.json').then(
-      (datasets: SeedDecisionDataset[]) => {
-        datasets = getDisplayedDatasets(datasets).filter(filter);
-
-        const countries = datasets
-          .flatMap((dataset) => dataset.routes)
-          .flatMap((route) => route.countries);
-
-        expect(rows).to.have.length(countries.length + 1);
-      },
+    cy.checkCsvDownload(
+      download,
+      `decisions-${format(new Date(), 'yyyyMMdd')}`,
+      filter,
     );
   });
-}
-
-function getDisplayedDatasets(
-  datasets: SeedDecisionDataset[],
-): SeedDecisionDataset[] {
-  const result = datasets.filter((dataset) =>
-    ['live', 'draft'].includes(dataset.status),
-  );
-
-  result.sort((dataset1, dataset2) => {
-    const professionComparison = dataset1.profession.localeCompare(
-      dataset2.profession,
-    );
-
-    if (professionComparison) {
-      return professionComparison;
-    }
-
-    const organisationComparison = dataset1.organisation.localeCompare(
-      dataset2.organisation,
-    );
-
-    if (organisationComparison) {
-      return organisationComparison;
-    }
-
-    return dataset2.year - dataset1.year;
-  });
-
-  return result;
-}
-
-function clickDownloadLink(link: string): void {
-  // This is a workaround for a Cypress bug to prevent it waiting
-  // indefinitely for a new page to load after clicking the download link
-  // See https://github.com/cypress-io/cypress/issues/14857
-  cy.window()
-    .document()
-    .then(function (doc) {
-      doc.addEventListener('click', () => {
-        setTimeout(function () {
-          doc.location.reload();
-        }, 5000);
-      });
-    });
-
-  cy.get('body').contains(link).click();
 }
